+++ src/Views/performance-dashboard.blade.php
@@ -12,7 +12,7 @@
             <h1 class="capitalize text-primary border-b-[2px] border-[var(--primary)] pl-20 py-5 text-2xl font-semibold mb-4">
                 {{ env('APP_NAME') }}
             </h1>
-            <div class="h-[80vh] w-full flex flex-col justify-between items-center gap-10 pl-4">
+            <div class="h-[80vh] w-full flex flex-col justify-between items-center gap-10 pl-4"  style="background: #ffffffc4; padding: 15px;">
                 <div class="content-wrapper w-full">
                     <h4 class="text-lg no-underline bg-primary text-white font-medium text-start px-6 py-3 mb-6 rounded-[4px] w-full">
                         Performance Dashboard